Copy improved appearance from Homepage to login
Improvements were previously made to the homepage (after login) to suit
various screen sizes.  The login page displays some of the same
introductory/explanatory information as the homepage, but this improved
formatting was not copied across to the login page.  This has now been
copied across - the homepage and login page are now consistent.