Add print to login_required
After transferring to Heroku, the app inconsistently throws the user
back to the login page.  It is unclear why this is happening but is
suspected to be due to user_id being returned as None (potentially
related to using SQLite instead of postgresql).  This print statement
will show in the log if this is in fact the cause of the error.