Re-engage memecache instead of filesystem cache
Both options (memecache and filesystem) are left in the code, with one
to be commented out.  Both remain in the code to allow easy swapping
between local testing and live operation on Heroku.  Local testing
cannot work with memecache as pylibmc does not seem to be compatible
with windows, and filesystem sessions is not compatible with Heroku.