WIP Improve appearance on mobile devices
This commit is Work In Progress, there are more visual improvements to
make.  This commit includes flask session options, so will not work
correctly on Heroku as this is still WIP.

- Messages changed to be collapsible to allow important content to be
visible with less scrolling).
- Adjusted padding around buttons and username/password boxes to better
suit mobile view and desktop view.
- Added username to 'load table' commentary.
- Removed 'Check' from homepage cube table, as unimportant to the user
most of the time (but retained on dedicated load page).
- Reduced padding of layout.html to use more screen space on small
devices.