Remove cube notation summary table
The intent behind the small table was to provide a quick reminder of
what cube notation meant, however the table would not look presentable
on all devices middle-size screens without taking up as much space as
the full description (at which point the summary table is wasteful).

The accordian button was also centered in the div.